+++ pkg/proxy/register.go
@@ -1,6 +1,7 @@
 package proxy
 
 import (
+	"fmt"
 	"net/http"
 	"net/http/httputil"
 	"net/url"
@@ -73,6 +74,9 @@ func (p *Register) Add(route *Route) error {
 }
 
 func (p *Register) createDirector(proxyDefinition *Definition, balancer Balancer) func(req *http.Request) {
+	paramNameExtractor := router.NewListenPathParamNameExtractor()
+	matcher := router.NewListenPathMatcher()
+
 	return func(req *http.Request) {
 		var upstreamURL string
 		if proxyDefinition.IsBalancerDefined() && balancer != nil {
@@ -106,7 +110,6 @@ func (p *Register) createDirector(proxyDefinition *Definition, balancer Balancer
 
 		if proxyDefinition.StripPath {
 			path = singleJoiningSlash(target.Path, req.URL.Path)
-			matcher := router.NewListenPathMatcher()
 			listenPath := matcher.Extract(proxyDefinition.ListenPath)
 
 			log.WithField("listen_path", listenPath).Debug("Stripping listen path")
@@ -114,6 +117,14 @@ func (p *Register) createDirector(proxyDefinition *Definition, balancer Balancer
 			if !strings.HasSuffix(target.Path, "/") && strings.HasSuffix(path, "/") {
 				path = path[:len(path)-1]
 			}
+		}
+
+		paramNames := paramNameExtractor.Extract(path)
+		parametrizedPath, err := p.applyParameters(req, path, paramNames)
+		if err != nil {
+			log.WithError(err).Warn("Unable to extract param from request")
+		} else {
+			path = parametrizedPath
 		}
 
 		log.WithField("path", path).Debug("Upstream Path")
@@ -150,6 +161,25 @@ func (p *Register) doRegister(listenPath string, handler http.HandlerFunc, metho
 			}
 		}
 	}
+}
+
+func (p *Register) applyParameters(req *http.Request, path string, paramNames []string) (string, error) {
+	for _, paramName := range paramNames {
+		paramValue := router.URLParam(req, paramName)
+
+		if len(paramValue) == 0 {
+			return "", errors.Errorf("unable to extract {%s} from request", paramName)
+		}
+
+		path = strings.Replace(
+			path,
+			fmt.Sprintf("{%s}", paramName),
+			paramValue,
+			-1,
+		)
+	}
+
+	return path, nil
 }
 
 func cleanSlashes(a string) string {